Breaking up files from __init__
Breaking up files to use inheritance.
