+++ check_mk_web_api/activate_mode.py
@@ -1,0 +1,10 @@
+class ActivateMode(enum.Enum):
+    """
+    # Members
+    DIRTY: Update sites with changes
+    ALL: Update all slave sites
+    SPECIFIC: Only update specified sites
+    """
+    DIRTY = 'dirty'
+    ALL = 'all'
+    SPECIFIC = 'specific'
