+++ check_mk_web_api/discover_mode.py
@@ -1,0 +1,12 @@
+class DiscoverMode(enum.Enum):
+    """
+    # Members
+    NEW: Only discover new services
+    REMOVE: Remove exceeding services
+    FIXALL: Remove exceeding services and discover new services (Tabula Rasa)
+    REFRESH: Start from scratch
+    """
+    NEW = 'new'
+    REMOVE = 'remove'
+    FIXALL = 'fixall'
+    REFRESH = 'refresh'
