+++ check_mk_web_api/web_api_servicegroup.py
@@ -1,0 +1,63 @@
+from check_mk_web_api.web_api_base import WebApiBase
+
+
+class WebApiServiceGroup(WebApiBase):
+
+    def get_servicegroup(self, group):
+        return self.get_all_servicegroups()[group]
+
+    def get_all_servicegroups(self):
+        """
+        Gets all service groups
+        """
+        return self.make_request('get_all_servicegroups')
+
+    def add_servicegroup(self, group, alias):
+        """
+        Adds a service group
+
+        # Arguments
+        group (str): name of group to add
+        alias (str): alias for group
+        """
+        data = NoNoneValueDict({
+            'groupname': group,
+            'alias': alias,
+        })
+
+        return self.make_request('add_servicegroup', data=data)
+
+    def edit_servicegroup(self, group, alias):
+        """
+        Edits a service group
+
+        # Arguments
+        group (str): name of group to edit
+        alias (str): new alias for group
+        """
+        data = NoNoneValueDict({
+            'groupname': group,
+            'alias': alias,
+        })
+
+        return self.make_request('edit_servicegroup', data=data)
+
+    def delete_servicegroup(self, group):
+        """
+        Deletes a service group
+
+        # Arguments
+        group (str): name of group to delete
+        """
+        data = NoNoneValueDict({
+            'groupname': group,
+        })
+
+        return self.make_request('delete_servicegroup', data=data)
+
+    def delete_all_servicegroups(self):
+        """
+        Deletes all service groups
+        """
+        for group in self.get_all_servicegroups():
+            self.delete_servicegroup(group)
